Small refactoring after code review and a few unit tests
