+++ src/common/Extensions/Base64Conversion/index.ts
@@ -1,0 +1,1 @@
+export * from "./InvocationArgument";
