+++ src/main/Extensions/Base64Conversion/Base64Converter.ts
@@ -1,0 +1,11 @@
+import { Buffer } from "buffer";
+
+export class Base64Converter {
+    public static encode(payload: string): string {
+        return Buffer.from(payload, "binary").toString("base64");
+    }
+
+    public static decode(payload: string): string {
+        return Buffer.from(payload, "base64").toString("binary");
+    }
+}
