+++ src/main/Extensions/Base64Conversion/index.ts
@@ -1,2 +1,1 @@
 export * from "./Base64Conversion";
-export * from "./InvocationArgument";
